requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
dnspython==2.4.2
gunicorn==21.2.0
dnspython==2.4.2
Flask==3.0.0
gunicorn==21.2.0
pymongo==4.6.1
pyrogram==2.0.106
python-dotenv==1.0.0
requests==2.31.0
TgCrypto==1.2.5
dnspython==2.4.2
Flask==3.0.0
gunicorn==21.2.0
pymongo==4.6.1
pyrogram==2.0.106
python-dotenv==1.0.0
requests==2.31.0
TgCrypto==1.2.5
